Clean user and car model dependencies in trip service

+++ carService/src/main/java/ru/itmo/carService/service/CarService.java
@@ -20,7 +20,7 @@ public class CarService {
         return repository.save(car);
     }
 
-    public Car getById(int id) {
+    public Car getById(long id) {
         return repository.findById(id).orElseThrow(IllegalArgumentException::new);
     }
